fix(types): Moves shared test function to a common file PE-11
